+++ packages/python-ta/CHANGELOG.md
@@ -28,6 +28,7 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Added `csv` response format to `RecursionTable`
 - Modified `SnapshotTracer` to trace into calls of functions defined in the same module that `SnapshotTracer` is called.
 - Added `webstepper_options` attribute to `SnapshotTracer` to allow configuration options for MemoryViz Webstepper, including the `line_context` option.
+- Modified PyTA CLI to optionally read file contents through standard input by passing the `--stdin` command-line argument or `-` as the filenames argument.
 
 ### 💫 New checkers